fix(feedback): harden the rating-clear path (#884 review round)
External review follow-ups: numerically normalize the clear sentinel so
a numeric-string "0" can't slip into the update/insert paths (validator
now also toInt()s), delete the full guest-scoped rating set on clear so
racy duplicate rows can't survive in the average (same defense as the
reaction path), and refresh the visible average/count after the
identity-modal submit path like the direct paths do.

+++ backend/src/utils/feedbackValidation.js
@@ -138,11 +138,13 @@ const validateFeedbackSubmission = [
     .withMessage('Invalid feedback type'),
 
   // Conditional validation based on feedback type. 0 clears the guest's
-  // existing rating (#884).
+  // existing rating (#884). toInt so a numeric string "0" reaches the
+  // service as a real 0 and hits the removal path.
   body('rating')
     .if(body('feedback_type').equals('rating'))
     .isInt({ min: 0, max: 5 })
-    .withMessage('Rating must be between 0 and 5'),
+    .withMessage('Rating must be between 0 and 5')
+    .toInt(),
 
   // Reactions (#839): fixed curated set only — no free-form emoji.
   body('reaction')
